api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy>=2.0
asyncpg
alembic
redis
python-jose[cryptography]
passlib[bcrypt]
motor
pandas
python-multipart
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
matplotlib==3.9.2
PyYAML==6.0.2
joblib==1.4.2
openpyxl==3.1.5
